Move subcommands into separate folder
